+++ app/src/main/java/co/kr/woowahan_repo/presentation/viewmodel/SignInViewModel.kt
@@ -5,8 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import co.kr.woowahan_repo.BuildConfig
-import co.kr.woowahan_repo.di.ServiceLocator
-import co.kr.woowahan_repo.domain.GithubTokenDataSource
+import co.kr.woowahan_repo.domain.datasource.GithubTokenDataSource
 import co.kr.woowahan_repo.domain.repository.GithubOAuthRepository
 import kotlinx.coroutines.launch
 import timber.log.Timber
